requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
<<<<<<< HEAD
Django==2.2.19
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.4
=======
Django==3.2.15
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
idna==3.4
isort==5.10.1
>>>>>>> branch_v2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
<<<<<<< HEAD
python3-openid==3.2.0
pytz==2022.6
=======
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
reportlab==3.6.12
>>>>>>> branch_v2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
